Project-NASA-Mission-Intelligence-Starter/requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
openai>=1.0.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0